feat(useEventContext): add autoOpenQuestionnaire option to control questionnaire display
The `autoOpenQuestionnaire` option has been added to the `useEventContext` hook to allow conditional display of the questionnaire based on the context. This provides more control over when the questionnaire is shown, improving user experience.

Additionally, the text in the `ChatHeader` component has been updated from "Nouveau" to "Nouvelle location" for better clarity.